# src/main/java/com/eaglebank/model/BadRequestErrorResponse.java
package com.eaglebank.model;

import java.util.List;

/**
 * Provides immutable API response data.
 *
 * @author mattbateup
 */
public record BadRequestErrorResponse(String message, List<FieldErrorDetail> details) {
}